Refactor hosting models into multiple modules
That hosting.py was getting quite unwieldy at over 1100 LOC, so i've had
a go at splitting it up: the external interface stays the same (as we
reexport everything used elsewhere in the codebase), but now the
"hosting" models are divided into a "datacentres" sub-package, a
"providers" sub-package, and an "abstract" sub-package for abbstract
base classes shared by both. The "providers" package is still pretty
huge, at 700 LOC, but I think this makes things a little more
manageable, with room for further improvement.

@mrchrisadams, what do you think? This is meant as a bit of a
conversation-starter as much as anything - would you like to see some of
the larger files split up, or are you happy with everything in one
place?
